requirements.txt
lancedb
langchain
langgraph
tavily-python
polygon
setuptools
ipykernel
langchain-openai
python-dotenv
langchain-community
fastapi
uvicorn
langchain-pinecone
langchain-tavily
python-multipart
jinja2
docx2txt
-e .